+++ mongo/reactive/src/integration/groovy/com/coditory/sherlock/ReactiveMongoDistributedLockSpec.groovy
@@ -1,10 +1,13 @@
 package com.coditory.sherlock
 
-class ReactiveMongoReleaseLockSpec extends ReleaseLockSpec implements UsesReactiveMongoSherlock {}
+class ReactiveMongoReleaseLockSpec extends ReleaseLockSpec
+        implements UsesReactiveMongoSherlock {}
 
-class ReactiveMongoAcquireLockSpec extends AcquireLockSpec implements UsesReactiveMongoSherlock {}
+class ReactiveMongoAcquireLockSpec extends AcquireLockSpec
+        implements UsesReactiveMongoSherlock {}
 
-class ReactiveMongoAcquireLockMultipleTimesSpec extends AcquireLockMultipleTimesSpec implements UsesReactiveMongoSherlock {
-}
+class ReactiveMongoAcquireLockMultipleTimesSpec extends AcquireLockMultipleTimesSpec
+        implements UsesReactiveMongoSherlock {}
 
-class ReactiveMongoInfiniteAcquireLockSpec extends InfiniteAcquireLockSpec implements UsesReactiveMongoSherlock {}
+class ReactiveMongoInfiniteAcquireLockSpec extends InfiniteAcquireLockSpec
+        implements UsesReactiveMongoSherlock {}