+++ mongo/sync/src/integration/groovy/com/coditory/sherlock/MongoDistributedLockSpec.groovy
@@ -1,9 +1,14 @@
 package com.coditory.sherlock
 
-class MongoReleaseLockSpec extends ReleaseLockSpec implements UsesMongoSherlock {}
+class MongoReleaseLockSpec extends ReleaseLockSpec
+        implements UsesMongoSherlock {}
 
-class MongoAcquireLockSpec extends AcquireLockSpec implements UsesMongoSherlock {}
+class MongoAcquireLockSpec extends AcquireLockSpec
+        implements UsesMongoSherlock {}
 
-class MongoAcquireLockMultipleTimesSpec extends AcquireLockMultipleTimesSpec implements UsesMongoSherlock {}
+class MongoAcquireLockMultipleTimesSpec extends AcquireLockMultipleTimesSpec
+        implements UsesMongoSherlock {}
 
-class MongoInfiniteAcquireLockSpec extends InfiniteAcquireLockSpec implements UsesMongoSherlock {}
+class MongoInfiniteAcquireLockSpec extends InfiniteAcquireLockSpec
+        implements UsesMongoSherlock {}
+