+++ mongo/sync/src/main/java/com/coditory/sherlock/MongoCollectionInitializer.java
@@ -13,6 +13,7 @@ class MongoCollectionInitializer {
     private final AtomicBoolean indexesCreated = new AtomicBoolean(false);
 
     MongoCollectionInitializer(MongoCollection<Document> collection) {
+        validateConnection(collection);
         this.collection = collection;
     }
 
@@ -23,5 +24,12 @@ class MongoCollectionInitializer {
         }
         collection.createIndex(INDEX, INDEX_OPTIONS);
         return collection;
+    }
+
+    private void validateConnection(MongoCollection<Document> collection) {
+        String readPreference = collection.getReadPreference().getName();
+        if (!"primary".equalsIgnoreCase(readPreference)) {
+            throw new IllegalArgumentException("Expected Mongo connection with readPreference=primary");
+        }
     }
 }