+++ sample/src/main/java/com/coditory/sherlock/sample/mongo/MongoSyncSample.java
@@ -1,0 +1,66 @@
+package com.coditory.sherlock.sample.mongo;
+
+import com.coditory.sherlock.DistributedLock;
+import com.coditory.sherlock.Sherlock;
+import com.coditory.sherlock.SherlockMigrator;
+import com.coditory.sherlock.sample.inmem.InMemSyncSample;
+import com.mongodb.client.MongoClient;
+import com.mongodb.client.MongoClients;
+import com.mongodb.client.MongoCollection;
+import org.bson.Document;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.time.Clock;
+import java.time.Duration;
+
+import static com.coditory.sherlock.MongoSherlockBuilder.mongoSherlock;
+
+public class MongoSyncSample {
+    private final Logger logger = LoggerFactory.getLogger(this.getClass());
+
+    private final Sherlock sherlock = mongoSherlock()
+            .withClock(Clock.systemDefaultZone())
+            .withLockDuration(Duration.ofMinutes(5))
+            .withUniqueOwnerId()
+            .withLocksCollection(locksCollection())
+            .build();
+
+    private static MongoCollection<Document> locksCollection() {
+        String database = "sherlock";
+        MongoClient mongoClient = MongoClients.create("mongodb://localhost:27017/" + database);
+        return mongoClient
+                .getDatabase("sherlock")
+                .getCollection("locks");
+    }
+
+    void sampleMongoLockUsage() {
+        logger.info(">>> SAMPLE: Lock usage");
+        DistributedLock lock = sherlock.createLock("sample-lock");
+        lock.acquireAndExecute(() -> logger.info("Lock acquired!"));
+    }
+
+    private void sampleInMemMigration() {
+        logger.info(">>> SAMPLE: Migration");
+        // first commit - all migrations are executed
+        new SherlockMigrator("db-migration", sherlock)
+                .addChangeSet("change set 1", () -> logger.info(">>> Change set 1"))
+                .addChangeSet("change set 2", () -> logger.info(">>> Change set 2"))
+                .migrate();
+        // second commit - only new change set is executed
+        new SherlockMigrator("db-migration", sherlock)
+                .addChangeSet("change set 1", () -> logger.info(">>> Change set 1"))
+                .addChangeSet("change set 2", () -> logger.info(">>> Change set 2"))
+                .addChangeSet("change set 3", () -> logger.info(">>> Change set 3"))
+                .migrate();
+    }
+
+    void runSamples() {
+        sampleMongoLockUsage();
+        sampleInMemMigration();
+    }
+
+    public static void main(String[] args) {
+        new MongoSyncSample().runSamples();
+    }
+}