+++ sql/src/main/java/com/coditory/sherlock/SqlQueries.java
@@ -39,12 +39,12 @@ final class SqlQueries {
 
     String deleteAcquiredByIdAndOwnerId() {
         return "DELETE FROM " + tableName
-                + " WHERE ID = ? AND ACQUIRED_BY = ? AND EXPIRES_AT > ?;";
+                + " WHERE ID = ? AND ACQUIRED_BY = ? AND (EXPIRES_AT IS NULL OR EXPIRES_AT > ?);";
     }
 
     String deleteAcquiredById() {
         return "DELETE FROM " + tableName
-                + " WHERE ID = ? AND EXPIRES_AT > ?;";
+                + " WHERE ID = ? AND (EXPIRES_AT IS NULL OR EXPIRES_AT > ?);";
     }
 
     String updateLockById() {