+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+## [0.4.12] - 2021-11-28
+### Changed
+- Fix releasing migration lock by SQL Connector [#61](https://github.com/coditory/sherlock-distributed-lock/issues/61)
+- Check MongoDb for valid readPreference
+
 ## [0.4.11] - 2021-11-24
 ### Changed
 - Update dependencies
@@ -20,7 +25,8 @@
 ### Changed
 - Fix documentation
 
-[Unreleased]: https://github.com/coditory/sherlock-distributed-lock/compare/v0.4.11...HEAD
+[Unreleased]: https://github.com/coditory/sherlock-distributed-lock/compare/v0.4.12...HEAD
+[0.4.12]: https://github.com/coditory/sherlock-distributed-lock/releases/tag/v0.4.12
 [0.4.11]: https://github.com/coditory/sherlock-distributed-lock/releases/tag/v0.4.11
 [0.4.10]: https://github.com/coditory/sherlock-distributed-lock/releases/tag/v0.4.10
 [0.4.9]: https://github.com/coditory/sherlock-distributed-lock/releases/tag/v0.4.9